+++ src/components/CallView/shared/VideoVue.spec.js
@@ -21,6 +21,7 @@
 
 import { createLocalVue, shallowMount } from '@vue/test-utils'
 import { cloneDeep } from 'lodash'
+import { createPinia, setActivePinia } from 'pinia'
 import Vuex from 'vuex'
 
 import VideoVue from './VideoVue.vue'
@@ -39,6 +40,7 @@ describe('VideoVue.vue', () => {
 	let localVue
 	let store
 	let testStoreConfig
+	let pinia
 
 	let callParticipantModel
 
@@ -87,6 +89,8 @@ describe('VideoVue.vue', () => {
 	beforeEach(() => {
 		localVue = createLocalVue()
 		localVue.use(Vuex)
+		pinia = createPinia()
+		setActivePinia(pinia)
 
 		testStoreConfig = cloneDeep(storeConfig)
 		// eslint-disable-next-line import/no-named-as-default-member
