+++ src/components/MediaSettings/MediaSettings.vue
@@ -208,8 +208,8 @@ import { CALL, VIRTUAL_BACKGROUND } from '../../constants.js'
 import { devices } from '../../mixins/devices.js'
 import isInLobby from '../../mixins/isInLobby.js'
 import BrowserStorage from '../../services/BrowserStorage.js'
+import { useGuestNameStore } from '../../stores/guestNameStore.js'
 import { localMediaModel } from '../../utils/webrtc/index.js'
-import { useGuestNameStore } from '../../store/guestNameStore.js'
 
 export default {
 	name: 'MediaSettings',
